+++ src/sdk/auth.js
@@ -1,11 +1,11 @@
 /*!
-* Copyright 2024-Present Animoca Brands Corporation Ltd. 
-* 
-* Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
-* 
-* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
-* 
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ * Copyright 2024-Present Animoca Brands Corporation Ltd.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 import AuthInfoManager from './lib/AuthInfoManager';
 import TokenManager from './lib/TokenManager';
@@ -13,10 +13,10 @@ import TransactionManager from './lib/TransactionManager';
 import { getStorageClass } from './lib/StorageManager';
 import { createPkceMeta, parseJwt, parseUrl, prepareTokenParams } from './utils';
 import { buildAuthEndpointUrl } from './endpoints';
-import { AuthError } from './utils/errors';
+import { AuthError, InvalidParamsError } from './utils/errors';
 
-export class OCAuthCore
-{
+export class OCAuthCore {
+    clientId;
     tokenManager;
     authInfoManager;
     transactionManager;
@@ -25,8 +25,10 @@ export class OCAuthCore
     logoutEndPoint;
     referralCode;
 
-    constructor ( loginEndpoint, redirectUri, transactionManager, tokenManager, referralCode, logoutEndPoint )
-    {
+    constructor(clientId, loginEndpoint, redirectUri, transactionManager, tokenManager, referralCode, logoutEndPoint) {
+        if (!clientId) {
+            throw new InvalidParamsError('clientId is not defined');
+        }
         this.transactionManager = transactionManager;
         this.tokenManager = tokenManager;
         this.authInfoManager = new AuthInfoManager();
@@ -34,17 +36,16 @@ export class OCAuthCore
         this.logoutEndPoint = logoutEndPoint;
         this.redirectUri = redirectUri;
         this.referralCode = referralCode;
+        this.clientId = clientId;
         this.syncAuthInfo();
     }
 
-    clearStorage ()
-    {
+    clearStorage() {
         this.transactionManager.clear();
         this.tokenManager.clear();
     }
 
-    async logout (logoutReturnTo)
-    {
+    async logout(logoutReturnTo) {
         this.clearStorage();
         const url = new URL(this.logoutEndPoint);
         if (logoutReturnTo) {
@@ -53,56 +54,49 @@ export class OCAuthCore
         window.location.assign(url.toString());
     }
 
-    async signInWithRedirect ( params )
-    {
+    async signInWithRedirect(params) {
         // we use ONLY code flow with PKCE, so lacks a lot of options
         // available in other OAuth SDKs.
-        const paramsClone = Object.assign( {}, params );
+        const paramsClone = Object.assign({}, params);
         paramsClone.redirectUri = this.redirectUri;
-        const signinParams = await prepareTokenParams( paramsClone );
-        const meta = createPkceMeta( signinParams );
-        this.transactionManager.save( meta );
+        paramsClone.clientId = this.clientId;
+        const signinParams = await prepareTokenParams(paramsClone);
+        const meta = createPkceMeta(signinParams);
+        this.transactionManager.save(meta);
         signinParams.referralCode = this.referralCode;
-        const requestUrl = buildAuthEndpointUrl( signinParams, this.loginEndPoint );
-        window.location.assign( requestUrl );
+        const requestUrl = buildAuthEndpointUrl(signinParams, this.loginEndPoint);
+        window.location.assign(requestUrl);
     }
 
-    async handleLoginRedirect ()
-    {
+    async handleLoginRedirect() {
         const urlParams = parseUrl();
         // Again we only handle PKCE code flow
-        if ( urlParams.code )
-        {
+        if (urlParams.code) {
             const meta = this.transactionManager.getTransactionMeta();
             const { codeVerifier } = meta;
-            if ( codeVerifier )
-            {
+            if (codeVerifier) {
                 // we used pkce mode, use it
-                await this.tokenManager.exchangeTokenFromCode( urlParams.code, codeVerifier, urlParams.state );
+                await this.tokenManager.exchangeTokenFromCode(urlParams.code, codeVerifier, urlParams.state);
                 // clear transaction meta, coz it's completed
                 this.transactionManager.clear();
                 this.syncAuthInfo();
                 return this.getAuthState();
-            } else
-            {
-                throw new AuthError( 'codeVerifier not found, cannot complete flow' );
+            } else {
+                throw new AuthError('codeVerifier not found, cannot complete flow');
             }
         }
 
-        // no code found, nothing to do 
+        // no code found, nothing to do
         return {};
     }
 
-    isAuthenticated ()
-    {
+    isAuthenticated() {
         // if both token exist and not expired
         return !this.tokenManager.hasExpired();
     }
 
-    syncAuthInfo ()
-    {
-        if ( this.tokenManager.hasExpired() )
-        {
+    syncAuthInfo() {
+        if (this.tokenManager.hasExpired()) {
             this.authInfoManager.clear();
         } else {
             const { edu_username, eth_address } = this.getParsedIdToken();
@@ -116,67 +110,57 @@ export class OCAuthCore
         }
     }
 
-    getAuthState ()
-    {
+    getAuthState() {
         return this.authInfoManager.getAuthState();
     }
 
-    getStateParameter ()
-    {
+    getStateParameter() {
         return this.tokenManager.getStateParameter();
     }
 
-    getIdToken ()
-    {
+    getIdToken() {
         return this.tokenManager.getIdToken();
     }
 
-    getAccessToken ()
-    {
+    getAccessToken() {
         return this.tokenManager.getAccessToken();
     }
 
-    getParsedIdToken ()
-    {
+    getParsedIdToken() {
         // return all info in id token
         const idToken = this.tokenManager.getIdToken();
-        if (idToken)
-        {
+        if (idToken) {
             return parseJwt(idToken);
         }
         return {};
     }
 
-    getParsedAccessToken ()
-    {
+    getParsedAccessToken() {
         // return all info in access token
         const accessToken = this.tokenManager.getAccessToken();
-        if (accessToken)
-        {
+        if (accessToken) {
             return parseJwt(accessToken);
         }
         return {};
     }
 
-    get OCId ()
-    {
+    get OCId() {
         const info = this.authInfoManager.getAuthState();
         return info.OCId ?? null;
     }
 
-    get ethAddress ()
-    {
+    get ethAddress() {
         const info = this.authInfoManager.getAuthState();
         return info.ethAddress ?? null;
     }
 }
 
-const LIVE_PUBLIC_KEY = 'MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEBIDHtLbgVM76SXZ4iuIjuO+ERQPnVpJzagOsZdYxFG3ZJmvfdpr/Z29SLUbdZWafrOlAVlKe1Ovf/tcH671tTw==';
-const SANDBOX_PUBLIC_KEY = 'MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE/EymMLXd/MVYPK5r2xXQj91ZVvX3OQ+QagvR2N6lCvRVjnzmOtPRTf+u5g1RliWnmuxbV3gTm0/0VuV/40Salg==';
-export class OCAuthLive extends OCAuthCore
-{
-    constructor ( opts = {} )
-    {
+const LIVE_PUBLIC_KEY =
+    'MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEBIDHtLbgVM76SXZ4iuIjuO+ERQPnVpJzagOsZdYxFG3ZJmvfdpr/Z29SLUbdZWafrOlAVlKe1Ovf/tcH671tTw==';
+const SANDBOX_PUBLIC_KEY =
+    'MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAE/EymMLXd/MVYPK5r2xXQj91ZVvX3OQ+QagvR2N6lCvRVjnzmOtPRTf+u5g1RliWnmuxbV3gTm0/0VuV/40Salg==';
+export class OCAuthLive extends OCAuthCore {
+    constructor(opts = {}) {
         const {
             tokenEndPoint: overrideTokenEndpoint,
             loginEndPoint: overrideLoginEndpoint,
@@ -184,6 +168,7 @@ export class OCAuthLive extends OCAuthCore
             publicKey: overridePublicKey,
             redirectUri,
             referralCode,
+            clientId,
         } = opts;
         const tokenEndpoint = overrideTokenEndpoint || 'https://api.login.opencampus.xyz/auth/token';
         const loginEndpoint = overrideLoginEndpoint || 'https://api.login.opencampus.xyz/auth/login';
@@ -191,16 +176,14 @@ export class OCAuthLive extends OCAuthCore
         const publicKey = overridePublicKey || LIVE_PUBLIC_KEY;
 
         const storageClass = getStorageClass(opts);
-        const pkceTransactionManager = new TransactionManager( storageClass );
-        const tokenManager = new TokenManager( storageClass, tokenEndpoint, publicKey );
-        super( loginEndpoint, redirectUri, pkceTransactionManager, tokenManager, referralCode, logoutEndpoint);
+        const pkceTransactionManager = new TransactionManager(storageClass);
+        const tokenManager = new TokenManager(storageClass, tokenEndpoint, publicKey);
+        super(clientId, loginEndpoint, redirectUri, pkceTransactionManager, tokenManager, referralCode, logoutEndpoint);
     }
 }
 
-export class OCAuthSandbox extends OCAuthCore
-{
-    constructor ( opts = {} )
-    {
+export class OCAuthSandbox extends OCAuthCore {
+    constructor(opts = {}) {
         const {
             tokenEndPoint: overrideTokenEndpoint,
             loginEndPoint: overrideLoginEndpoint,
@@ -209,14 +192,15 @@ export class OCAuthSandbox extends OCAuthCore
             redirectUri,
             referralCode,
         } = opts;
+        const clientId = opts.clientId || 'sandbox';
         const tokenEndpoint = overrideTokenEndpoint || 'https://api.login.sandbox.opencampus.xyz/auth/token';
         const loginEndpoint = overrideLoginEndpoint || 'https://api.login.sandbox.opencampus.xyz/auth/login';
         const logoutEndpoint = overrideLogoutEndpoint || 'https://api.login.sandbox.opencampus.xyz/auth/logout';
         const publicKey = overridePublicKey || SANDBOX_PUBLIC_KEY;
 
         const storageClass = getStorageClass(opts);
-        const pkceTransactionManager = new TransactionManager( storageClass );
-        const tokenManager = new TokenManager( storageClass, tokenEndpoint, publicKey );
-        super( loginEndpoint, redirectUri, pkceTransactionManager, tokenManager, referralCode, logoutEndpoint);
+        const pkceTransactionManager = new TransactionManager(storageClass);
+        const tokenManager = new TokenManager(storageClass, tokenEndpoint, publicKey);
+        super(clientId, loginEndpoint, redirectUri, pkceTransactionManager, tokenManager, referralCode, logoutEndpoint);
     }
 }